+++ docs/src/index.md
@@ -109,6 +109,7 @@ RunAt
 ParameterSource
 Aux
 Grid
+Delay
 ```
 
 ### Custom Rule interface and helpers
