+++ src/flags.jl
@@ -150,69 +150,3 @@ output = sim!(output, rule; boundary=Remove())
 ```
 """
 struct Remove <: BoundaryCondition end
-
-
-"""
-    ParameterSource
-
-Abstract supertypes for parameter source wrappers. These allow
-parameters to be retreived from auxilliary data or from other grids.
-"""
-abstract type ParameterSource end
-
-"""
-    Aux <: ParameterSource
-
-    Aux{K}()
-    Aux(K::Symbol)
-
-Use auxilary array with key `K` as a parameter source.
-
-Implemented in rules with:
-
-```julia
-get(data, rule.myparam, index...)
-```
-
-When an `Aux` param is specified at rule construction with:
-
-```julia
-rule = SomeRule(; myparam=Aux{:myaux})
-output = ArrayOutput(init; aux=(myaux=myauxarray,))
-```
-
-If the array is a DimensionalData.jl `DimArray` with a `Ti` (time)
-dimension, the correct interval will be selected automatically,
-precalculated for each timestep so it has no significant overhead.
-"""
-struct Aux{K} end
-Aux(key::Symbol) = Aux{key}()
-
-_unwrap(::Aux{X}) where X = X
-_unwrap(::Type{<:Aux{X}}) where X = X
-
-"""
-    Grid <: ParameterSource
-
-    Grid{K}()
-    Grid(K::Symbol)
-
-Use grid with key `K` as a parameter source.
-
-Implemented in rules with:
-
-```julia
-get(data, rule.myparam, index...)
-```
-
-And specified at rule construction with:
-
-```julia
-SomeRule(; myparam=Grid{:somegrid})
-```
-"""
-struct Grid{K} end
-Grid(key::Symbol) = Grid{key}()
-
-_unwrap(::Grid{X}) where X = X
-_unwrap(::Type{<:Grid{X}}) where X = X
